(rename) rename EmailForm to StartProcessingForm (reference https://github.com/AlexsLemonade/refinebio-web/pull/202\#issuecomment-1767050652)
